# api/docs/swagger.yaml
definitions:
...
  user.AddEcoPointsRequest:
    properties:
      points:
        type: integer
      reason:
        type: string
      user_id:
        type: string
    type: object
  user.AddEcoPointsResponse:
    properties:
      added_points:
        type: integer
      eco_points:
        type: integer
      reason:
        type: string
      timestamp:
        type: string
      user_id:
        type: string
    type: object
  user.EcoPointTransaction:
    properties:
      id:
        type: string
      points:
        type: integer
      reason:
        type: string
      timestamp:
        type: string
      type:
        type: string
    type: object
  user.GetEcoPointsHistoryRequest:
    properties:
      limit:
        type: integer
      page:
        type: integer
      points:
        type: integer
      reason:
        type: string
      type:
        type: string
      user_id:
        type: string
    type: object
  user.GetEcoPointsHistoryResponse:
    properties:
      history:
        items:
          $ref: '#/definitions/user.EcoPointTransaction'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
  user.GetEcoPointsResponse:
    properties:
      eco_points:
        type: integer
      last_updated:
        type: string
      user_id:
        type: string
    type: object
  user.GetUserProfileResponse:
    properties:
      created_at:
        type: string
      eco_points:
        type: integer
      email:
        type: string
      full_name:
        type: string
      id:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
  user.GetUsersRequest:
    properties:
      eco_points:
        type: integer
      full_name:
        type: string
      limit:
        type: integer
      page:
        type: integer
      username:
        type: string
    type: object
  user.GetUsersResponse:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      users:
        items:
          $ref: '#/definitions/user.User'
        type: array
    type: object
  user.UpdateProfileResponse:
    properties:
      bio:
        type: string
      email:
        type: string
      full_name:
        type: string
      id:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
  user.UpdateUserProfileRequest:
    properties:
      bio:
        type: string
      full_name:
        type: string
      user_id:
        type: string
      username:
        type: string
    type: object
  user.User:
    properties:
      eco_points:
        type: integer
      full_name:
        type: string
      id:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API Gateway of User Item System
  title: User Item System
  version: "1.0"
paths:
...
  /item-system/users:
    post:
      description: Retrieves list of users from PostgreSQL
      parameters:
      - description: filter user info
        in: body
        name: new_info
        required: true
        schema:
          $ref: '#/definitions/user.GetUsersRequest'
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/user.GetUsersResponse'
            type: array
        "500":
          description: Server error while getting users
          schema:
            type: string
      summary: Gets list of users
      tags:
      - user
  /item-system/users/{user_id}:
    delete:
      description: Removes user info from PostgreSQL
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      responses:
        "200":
          description: User deleted successfully
          schema:
            type: string
        "400":
          description: Invalid user ID
          schema:
            type: string
        "500":
          description: Server error while deleting user
          schema:
            type: string
      summary: Deletes a user
      tags:
      - user
    get:
      description: Retrieves user profile info from PostgreSQL
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.GetUserProfileResponse'
        "400":
          description: Invalid user ID
          schema:
            type: string
        "500":
          description: Server error while getting user profile
          schema:
            type: string
      summary: Gets user profile
      tags:
      - user
    put:
      description: Updates user profile info in PostgreSQL
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      - description: Update user info
        in: body
        name: new_info
        required: true
        schema:
          $ref: '#/definitions/user.UpdateUserProfileRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.UpdateProfileResponse'
        "400":
          description: Invalid user ID or data
          schema:
            type: string
        "500":
          description: Server error while updating user profile
          schema:
            type: string
      summary: Updates user profile
      tags:
      - user
  /item-system/users/{user_id}/eco-points:
    get:
      description: Retrieves eco points info from PostgreSQL
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.GetEcoPointsResponse'
        "400":
          description: Invalid user ID
          schema:
            type: string
        "500":
          description: Server error while getting eco points
          schema:
            type: string
      summary: Gets eco points of a user
      tags:
      - user
    put:
      description: Inserts eco points info into PostgreSQL
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      - description: Eco points info
        in: body
        name: points
        required: true
        schema:
          $ref: '#/definitions/user.AddEcoPointsRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.AddEcoPointsResponse'
        "400":
          description: Invalid user ID or data
          schema:
            type: string
        "500":
          description: Server error while adding eco points
          schema:
            type: string
      summary: Adds eco points to a user
      tags:
      - user
  /item-system/users/{user_id}/eco-points/history:
    post:
      description: Retrieves eco points history from PostgreSQL
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      - description: Eco points history info
        in: body
        name: points
        required: true
        schema:
          $ref: '#/definitions/user.GetEcoPointsHistoryRequest'
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/user.GetEcoPointsHistoryResponse'
            type: array
        "400":
          description: Invalid user ID
          schema:
            type: string
        "500":
          description: Server error while getting eco points history
          schema:
            type: string
      summary: Gets eco points history of a user
      tags:
      - user
swagger: "2.0"